FEAT : (#14) 분실신고 목록 조회 api 생성

+++ src/main/java/com/api/goldentime/web/dto/response/post/LostPostSaveResponseDto.java
@@ -6,12 +6,12 @@ import lombok.Getter;
 @Getter
 public class LostPostSaveResponseDto {
 
-    private final int id;
+    private final Long id;
     private final String kind;
     private final String color;
 
     @Builder
-    public LostPostSaveResponseDto(int id, String kind, String color) {
+    public LostPostSaveResponseDto(Long id, String kind, String color) {
         this.id = id;
         this.kind = kind;
         this.color = color;
